pyproject.toml
[tool.poetry]
name = "sporthive-api-scraper"
version = "0.1.0"
description = ""
authors = ["Florian Maas <fpgmaas@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
numpy = "^1.22.3"
requests = "^2.27.1"
pandas = "^1.4.2"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]
black = "^22.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
